+++ Marvel/Core/UseCases/CharacterUseCase.swift
@@ -8,15 +8,9 @@
 
 import Foundation
 
-// MARK: - Typealias
+// MARK: - Paginator
 
 typealias CharacterPaginator = Paginator<MarvelCharacter>
-
-// MARK: - UseCase
-
-protocol CharacterUseCase {
-    func loadCharacters(with parameter: CharacterParameter) -> Promise<CharacterPaginator>
-}
 
 // MARK: - Parameters
 
@@ -33,5 +27,30 @@ struct CharacterParameter: Parameter {
         self.nameStartsWith = query
         self.limit = limit
         self.offset = offset
+    }
+}
+
+// MARK: - UseCase
+
+protocol CharacterUseCase {
+    func loadCharacters(with parameter: CharacterParameter) -> Promise<CharacterPaginator>
+}
+
+final class DefaultCharacterUseCase {
+    private let gateway: CharacterGateway
+    private let repository: AnyRepository<MarvelCharacter>
+
+    init(gateway: CharacterGateway, repository: AnyRepository<MarvelCharacter>) {
+        self.gateway = gateway
+        self.repository = repository
+    }
+}
+
+extension DefaultCharacterUseCase: CharacterUseCase {
+    func loadCharacters(with parameter: CharacterParameter) -> Promise<CharacterPaginator> {
+        gateway
+            .loadCharacters(with: .init(parameter))
+            .then { self.repository.save(entites: $0.results) }
+            .recover { _ in self.repository.fetchAll().then { CharacterPaginator(results: $0) } }
     }
 }
